Add files via upload
Overview
The 16 MBTI Types project is a web application designed to help users understand the 16 personality types based on the Myers-Briggs Type Indicator (MBTI). The application provides an interactive and visually appealing user experience with detailed information on each personality type, as well as charts and graphs to visualize the results.

This project is built as a full-stack web application using Django for the backend and Tailwind CSS for styling. Additionally, D3.js is used for creating interactive and dynamic charts to display personality data in a visually appealing way.